pyproject.toml
[tool.poetry]
name = "django-shop"
version = "0.1.0"
description = ""
authors = ["macalistervadim <macalistervadim@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
python-dotenv = "^1.0.1"
pillow = "^10.0.0"
celery = "^5.4.0"
flower = "^2.0.1"
stripe = "^11.2.0"
psycopg2-binary = "^2.9.10"
weasyprint = "^63.0"
redis = "^5.2.1"
django-rosetta = "^0.10.1"
django-parler = "^2.3"
django-localflavor = "^4.0"

[tool.poetry.group.dev.dependencies]
flake8-import-order = "^0.18.2"
flake8-print = "^5.0.0"
flake8-quotes = "^3.4.0"
flake8-use-pathlib = "^0.3.0"
pep8-naming = "^0.14.1"
flake8 = "^7.1.1"
flake8-absolute-import = "^1.0.0.2"
flake8-commas = "^4.0.0"
black = "^24.10.0"
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=79
